[fix] type

+++ AccountingScholarships.Domain/Entities/epvosso/SpecialitiesEpvo.cs
@@ -9,7 +9,7 @@ namespace AccountingScholarships.Domain.Entities.epvosso
     public class SpecialitiesEpvo
     {
         public double? UniversityId { get; set; }
-        public float? Id { get; set; }
+        public double? Id { get; set; }
         public double? profCafId { get; set; }
         public string? NameRu { get; set; }
         public string? NameKz { get; set; }
